"
wrap around on 2038-01-19 03:14:07 UTC.  Such dates are valid deletion
times starting 2018-01-19 with the 20 years long maximum ttl.

This patchset extends gc_clock::duration::rep to int64_t and adds
respective unit tests for the max_ttl cases.

Fixes #3353

Tests: unit (release)
"

* 'projects/gc_clock_64/v2' of https://github.com/bhalevy/scylla:
  tests: cql_query_test add test_time_overflow
  gc_clock: make 64 bit
  sstables: mc: use int64_t for local_deletion_time and ttl
  sstables: add capped_tombstone_deletion_time stats counter
  sstables: mc: cap partition tombstone local_deletion_time to max
  sstables: add capped_local_deletion_time stats counter
  sstables: mc: metadata collector: cap local_deletion_time at max
  sstables: mc: use proper gc_clock types for local_deletion_time and ttl
  db: get default_time_to_live as int32_t rather than gc_clock::rep
  sstables: safely convert ttl and local_deletion_time to int32_t
  sstables: mc: move liveness_info initialization to members
  sstables: mc: move parsing of liveness_info deltas to data_consume_rows_context_m
  sstables: mc: define expired_liveness_ttl as signed int32_t
  sstables: mc: change write_delta_deletion_time to receive tombstone rather than deletion_time
  sstables: mc: use gc_clock types for writing delta ttl and local_deletion_time

+++ tests/sstable_test.cc
@@ -328,7 +328,7 @@ public:
     }
 
     virtual proceed consume_cell(bytes_view col_name, bytes_view value,
-            int64_t timestamp, int32_t ttl, int32_t expiration) override {
+            int64_t timestamp, int64_t ttl, int64_t expiration) override {
         BOOST_REQUIRE(ttl == 0);
         BOOST_REQUIRE(expiration == 0);
         switch (count_cell) {
@@ -444,7 +444,7 @@ public:
         return proceed::yes;
     }
     virtual proceed consume_cell(bytes_view col_name, bytes_view value,
-            int64_t timestamp, int32_t ttl, int32_t expiration) override {
+            int64_t timestamp, int64_t ttl, int64_t expiration) override {
         count_cell++;
         return proceed::yes;
     }
@@ -596,7 +596,7 @@ public:
     }
 
     virtual proceed consume_cell(bytes_view col_name, bytes_view value,
-            int64_t timestamp, int32_t ttl, int32_t expiration) override {
+            int64_t timestamp, int64_t ttl, int64_t expiration) override {
         switch (count_cell) {
         case 0:
             // The silly "cql row marker" cell
